Remove build directory only if it exists, makefile

+++ makefile
@@ -86,5 +86,5 @@ $(DEBUG_O): $(DEBUG_C) $(DEBUG_O_HEADERS)
 	$(COBJ) $@ $<
 
 clean:
-	rm -r $(BUILD_DIR)
+	if [ -d "$(BUILD_DIR)" ]; then rm -r $(BUILD_DIR); fi
